requirements.txt
# requirements.txt

# --- Core Application & LangChain Ecosystem ---
# Pinned to the EXACT versions from your original working environment to guarantee stability.
langchain==0.3.25
langchain-core==0.3.65
langgraph==0.4.8
langchain-google-genai==2.1.5
# **THE FIX**: Using the correct package name and version
google-genai==1.19.0

# --- Vector Database ---
pymilvus==2.5.11

# --- Data Handling & Utilities ---
pandas==2.3.0
python-dotenv==1.1.0
pydantic==2.11.5
tqdm==4.67.1

# --- Marshmallow Pin for a Pymilvus Sub-dependency ---
# This is still required based on the libraries in your working environment.
marshmallow<3.0.0

streamlit